requirements.txt
feedparser==6.0.11
anthropic>=0.40.0
jinja2>=3.1.4
python-dateutil>=2.9.0
markdown>=3.6
